lesson_01/requirements.txt
pyjwt
requests
streamlit
zhipuai
python-dotenv